+++ src/main/java/edu/city/studentuml/model/domain/Dependency.java
@@ -8,25 +8,25 @@ import java.io.Serializable;
 import org.w3c.dom.Element;
 
 /**
- * 
  * @author Ervin Ramollari
  */
 public class Dependency implements Serializable, IXMLCustomStreamable {
 
     private DesignClass from;
     private DesignClass to;
+    private String stereotype;
 
     public Dependency(DesignClass a, DesignClass b) {
         from = a;
         to = b;
+        stereotype = null;
     }
-
 
     /*
      * DO NOT CHANGE THE NAME: CALLED BY REFLECTION IN CONSISTENCY CHECK
      *
      * if name is changed the rules.txt / file needs to be updated
-     */    
+     */
     public DesignClass getFrom() {
         return from;
     }
@@ -35,16 +35,31 @@ public class Dependency implements Serializable, IXMLCustomStreamable {
      * DO NOT CHANGE THE NAME: CALLED BY REFLECTION IN CONSISTENCY CHECK
      *
      * if name is changed the rules.txt / file needs to be updated
-     */    
+     */
     public DesignClass getTo() {
         return to;
     }
 
+    public String getStereotype() {
+        return stereotype;
+    }
+
+    public void setStereotype(String stereotype) {
+        this.stereotype = stereotype;
+    }
+
     public void streamFromXML(Element node, XMLStreamer streamer, Object instance) {
+        String stereo = node.getAttribute("stereotype");
+        if (stereo != null && !stereo.isEmpty()) {
+            stereotype = stereo;
+        }
     }
 
     public void streamToXML(Element node, XMLStreamer streamer) {
         node.setAttribute("from", SystemWideObjectNamePool.getInstance().getNameForObject(from));
         node.setAttribute("to", SystemWideObjectNamePool.getInstance().getNameForObject(to));
+        if (stereotype != null && !stereotype.isEmpty()) {
+            node.setAttribute("stereotype", stereotype);
+        }
     }
 }